AI-scoring-model/api/requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ML dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=9.0.0
numpy>=1.21.0

# Pydantic for data validation
pydantic>=2.0.0
